+++ .vscode/launch.json
@@ -5,11 +5,11 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Python: Current File",
+            "name": "pr-report",
             "type": "python",
             "request": "launch",
-            "program": "${file}",
-            "console": "integratedTerminal"
+            "program": "${workspaceFolder}/.venv/bin/pr-report",
+            "console": "integratedTerminal",
         }
     ]
 }
